changemnent de la vitesse

+++ UI_VIBECODE/main.py
@@ -211,8 +211,8 @@ class CameraThread(QThread):
                                 label += " [LIMITE]"
 
                             label += f" | X={Xcm:+.1f} Y={Ycm:+.1f}"
-                            angle_deg = det.get("angle", 0.0)   # si ton détecteur fournit un angle
-                            current_blocks.append([col, "2x4", round(-Xcm, 2), round(Ycm, 2), float(self.z_table), angle_deg])
+                            angle_deg = float(det.get("angle", 0.0))
+                            current_blocks.append([col, "2x4", float(round(-Xcm, 2)), float(round(Ycm, 2)), float(self.z_table), angle_deg])
 
 
                     cv2.putText(rgb, label, (int(cx) + 8, int(cy) - 8),
